Apply the odf-converter-1.1-load-wordconverter.diff patch from Novell's build of odf-converter:
This patch fixes source/Common/OdfZipUtils/ZipArchiveWriter.cs to use
Path.DirectorySeparatorChar instead of hardcoding "\\". Also, in a
"static" build, "Static" is inserted in the assembly name, as is the
convention in the Makefiles for building with Mono for mkbundle'ing.

Then source/Word/Converter/Makefile is changed so that the name of the
generated assembly is WordConverter.dll and not Word.dll, to match what is
generated by the .csproj file, and expected by
source/Word/Converter/resources/odf2oox/2oox-part_relationships.xsl and
source/Word/Converter/resources/oox2odf/2odf-ole.xsl. (Yeah, really nice to
have such a coupling from a .xsl file to the name of the assembly...) The
generated "static" assembly is correspondingly called WordConverterStatic.dll.

Then in source/Shell/OdfConverterTest/Makefile the name of the Word assemblies
are changed accordingly.



git-svn-id: http://svn.code.sf.net/p/odf-converter/code/trunk@4263 24c45a59-3718-0410-950f-d1d872dd8675