Apply odf-converter-1.0.6-value-of.diff from Novell's build of odf-converter:
Mono's XSLT implementation doesn't like separate start and end tags
for xsl:value-of, even if only whitescape and/or XML comments is
inbetween.

The DTD provided with the XSLT standard does also say thay the
xsl:value-of should be EMPTY, but that is not normative. The XSL
standard says that empty element tags SHOULD be used (and only used)
for EMPTY elements. "SHOULD", not "MUST", so Mono probably is too
picky here.



git-svn-id: http://svn.code.sf.net/p/odf-converter/code/trunk@4266 24c45a59-3718-0410-950f-d1d872dd8675