Fix for #2780337: KB908002 did not get installed if it was already installed for a different user on that machine. The checks in the bootstrapped package have been modified to check whether KB908002 is actually installed for the current user.
Installation state	Return integer value	Meaning
INSTALLSTATE_INVALIDARG	-2	An invalid parameter was passed to the function.
INSTALLSTATE_UNKNOWN	-1	The product is not advertised or installed.
INSTALLSTATE_ADVERTISED	1	The product is advertised but not installed.
INSTALLSTATE_ABSENT	2	The product is installed for a different user.
INSTALLSTATE_DEFAULT	5	The product is installed for the current user.

git-svn-id: http://svn.code.sf.net/p/odf-converter/code/trunk@5240 24c45a59-3718-0410-950f-d1d872dd8675