The two C# snippets embedded in the XSL have changed, so the C# source files extracted from those for pre-compilation had to be updated correspondingly.
This is a rather tediouos task, as it involves running the mkbundled
odfconverter with
MONO_BUNDLED_OPTIONS='--trace=M:System.Reflection.Assembly:LoadFrom'
to see the name of the assembly it tries to load (like for instance
Scripte41bc1aee0566b95b29442ba7e2f770f.dll), and then create the
corresponding
source/Shell/OdfConverter/Scripte41bc1aee0566b95b29442ba7e2f770f.cs,
put the extracted C# snippet in it, and modify
source/Shell/OdfConverter/Makefile correspondingly. It would be nice
to automate this somehow, but I dunno, it's not often we need to do
this.



git-svn-id: http://svn.code.sf.net/p/odf-converter/code/trunk@5272 24c45a59-3718-0410-950f-d1d872dd8675